Updated forms for a better UX
Added an authentication check with the different forms to prevent anonymous
users from executing DB commands. The forms simply will not display if a user
is not signed in. If the URL for a form is manually typed, the authentication
check runs before the form will display.

Improved the user experience by adding an onClick feature to the overlay
of each of the app forms. This will allow the user to close the form by
clicking on the overlay.